Update package.json
The key order is now more aligned to the package.json documentation for convenience. Extra data is added (can easily be removed if not needed). `private` added since most themes won't want to be added to NPM directly. `theme` options added, since some of these values (like name and description) will be visible to the end client, and may not represent what a developer should see.