Move config file
Moving it to a top-level directory of `/config/` makes it easier to find (like `functions.php`). Putting it as one file in a directory means other files could be added to the same directory. It also happens to be the default location that BrightNucleus/Config package looks for config files.