Rename lib to src
While Genesis uses a `lib/` directory for the bulk of it's source code, there is no requirement or a hard expectation that child themes must do the same. As such, a far more common naming choice is `src/`, since this is used over many more PHP packages to indicate where the source code (as opposed to test or other code) lives. The code inside was never really library code (often understood to mean an external bit of code written by others), so renaming to `src/` removes this potential ambiguity.